Move pif and n64_cic_nus_6105 modules into the SI subsystem.
