+++ projects/msvc11/mupen64plus-core.vcxproj
@@ -152,11 +152,11 @@
     <ClCompile Include="..\..\src\main\md5.c" />
     <ClCompile Include="..\..\src\memory\memory.c" />
     <ClCompile Include="..\..\src\r4300\mi_controller.c" />
-    <ClCompile Include="..\..\src\memory\n64_cic_nus_6105.c" />
+    <ClCompile Include="..\..\src\si\n64_cic_nus_6105.c" />
     <ClCompile Include="..\..\src\osd\OGLFT.cpp" />
     <ClCompile Include="..\..\src\osd\osd.cpp" />
     <ClCompile Include="..\..\src\pi\pi_controller.c" />
-    <ClCompile Include="..\..\src\memory\pif.c" />
+    <ClCompile Include="..\..\src\si\pif.c" />
     <ClCompile Include="..\..\src\plugin\plugin.c" />
     <ClCompile Include="..\..\src\main\profile.c" />
     <ClCompile Include="..\..\src\r4300\pure_interp.c" />
@@ -229,12 +229,12 @@
     <ClInclude Include="..\..\src\main\md5.h" />
     <ClInclude Include="..\..\src\memory\memory.h" />
     <ClInclude Include="..\..\src\r4300\mi_controller.h" />
-    <ClInclude Include="..\..\src\memory\n64_cic_nus_6105.h" />
+    <ClInclude Include="..\..\src\si\n64_cic_nus_6105.h" />
     <ClInclude Include="..\..\src\osd\OGLFT.h" />
     <ClInclude Include="..\..\src\r4300\ops.h" />
     <ClInclude Include="..\..\src\osd\osd.h" />
     <ClInclude Include="..\..\src\pi\pi_controller.h" />
-    <ClInclude Include="..\..\src\memory\pif.h" />
+    <ClInclude Include="..\..\src\si\pif.h" />
     <ClInclude Include="..\..\src\plugin\plugin.h" />
     <ClInclude Include="..\..\src\osal\preproc.h" />
     <ClInclude Include="..\..\src\main\profile.h" />
